chore: organise MUI theme component overrides
