+++ site/src/theme/components/dataDisplay/MuiListItemTextThemeOptions.ts
@@ -1,5 +1,5 @@
 import { Components } from "@mui/material";
-import { customColors } from "../palette";
+import { customColors } from "../../palette";
 
 export const MuiListItemTextThemeOptions: Components["MuiListItemText"] = {
   styleOverrides: {
