+++ site/src/theme/components/feedback/MuiSkeletonThemeOptions.ts
@@ -1,5 +1,5 @@
 import { Components } from "@mui/material";
-import { customColors } from "../palette";
+import { customColors } from "../../palette";
 
 export const MuiSkeletonThemeOptions: Components["MuiSkeleton"] = {
   styleOverrides: {
