+++ site/src/theme/components/inputs/MuiInputBaseThemeOptions.ts
@@ -1,5 +1,5 @@
 import { Components } from "@mui/material";
-import { customColors } from "../palette";
+import { customColors } from "../../palette";
 
 export const MuiInputBaseThemeOptions: Components["MuiInputBase"] = {
   styleOverrides: {
