+++ site/src/theme/components/navigation/MuiLinkThemeOptions.ts
@@ -1,5 +1,5 @@
 import { Components } from "@mui/material";
-import { customColors } from "../palette";
+import { customColors } from "../../palette";
 
 export const MuiLinkThemeOptions: Components["MuiLink"] = {
   styleOverrides: {
