+++ site/src/theme/components/navigation/MuiTabThemeOptions.ts
@@ -1,6 +1,6 @@
 import { Components } from "@mui/material";
-import { customColors } from "../palette";
-import { defaultTheme } from "../util";
+import { customColors } from "../../palette";
+import { defaultTheme } from "../../util";
 
 export const MuiTabThemeOptions: Components["MuiTab"] = {
   defaultProps: {
