+++ site/src/theme/components/surfaces/MuiPaperThemeOptions.ts
@@ -1,5 +1,5 @@
 import { Components } from "@mui/material";
-import { customColors } from "../palette";
+import { customColors } from "../../palette";
 
 export const MuiPaperThemeOptions: Components["MuiPaper"] = {
   styleOverrides: {
